feat(uwsm): restart Waybar using its systemd service if enabled (#46)
* Restart Waybar using its systemd service if it's enabled

For uwsm users, Hyprland recommends enabling Waybar via its Systemd service. The script will now check if the user is running Waybar using the Systemd service or not, if the service is not enabled, the script will restart Waybar like usual.

* Refactor Waybar restart logic in install.sh

Remove unnecessary check for the existence of the Waybar Systemd service.

* format

* Check if the service is enabled and restarted

* format

---------

+++ install.sh
@@ -47,9 +47,12 @@ main() {
 
 	cprintf blue "\nRestarting Waybar..."
 
-	pkill waybar
-	waybar &> /dev/null &
-	disown
+	if ! (systemctl --user is-enabled waybar.service &&
+		systemctl --user restart waybar.service) &> /dev/null; then
+		pkill waybar
+		waybar &> /dev/null &
+		disown
+	fi
 
 	if ((ERRORS > 0)); then
 		cprintf red "\nInstallation completed with $ERRORS errors"
